+++ pages/_app.tsx
@@ -1,7 +1,13 @@
+import { Provider } from 'react-redux'
+import { store } from 'Stores/stores'
 import '../styles/globals.css'
 
-function MyApp({ Component, pageProps }) {
-  return <Component {...pageProps} />
+export function MyApp({ Component, pageProps }) {
+  return (
+    <Provider store={store}>
+      <Component {...pageProps} />
+    </Provider>
+  )
 }
 
 export default MyApp
